Move enums from @types to @symbols
