+++ packages/symbols/src/index.ts
@@ -8,5 +8,7 @@ export { SequenceMessageCode } from "./sequence-status-code";
 export { OpRecordCode } from "./op-record-code";
 export { APIErrorCode } from "./api-error-codes";
 export { DisconnectHubErrors } from "./disconnect-error-codes";
+export { InstanceStatus } from "./instance-status";
 
+export * from "./sd-stream-handler-state";
 export * from "./headers";
