+++ packages/symbols/src/instance-status.ts
@@ -1,0 +1,10 @@
+
+export const enum InstanceStatus {
+    INITIALIZING = "initializing",
+    STARTING = "starting",
+    RUNNING = "running",
+    STOPPING = "stopping",
+    KILLING = "killing",
+    COMPLETED = "completed",
+    ERRORED = "errored",
+}
