+++ packages/types/src/sd-stream-handler.ts
@@ -1,28 +1,6 @@
-export enum WorkState {
-    Flowing = "flowing",
-    Close = "close",
-    Error = "error"
-}
-
-export enum WritableState {
-    Finish = "finish",
-    Writable = "writable",
-    Drain = "drain"
-}
-
-export enum ReadableState {
-    Readable = "readable",
-    Pause = "pause",
-    End = "end"
-}
+import { WorkState, WritableState, ReadableState, StreamType } from "@scramjet/symbols";
 
 export type StreamState = WorkState | WritableState | ReadableState;
-
-export enum StreamType {
-    Instance = "instance",
-    Topic = "topic",
-    Api = "api"
-}
 
 export type StreamOptions = Record<string, any>;
